requirements.txt
sqlalchemy==2.0.23
pandas==2.1.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9
streamlit==1.29.0
pytz==2023.3
requests==2.31.0 